+++ Nextron/thor_cloud_modules/__init__.py
@@ -1,5 +1,6 @@
 from sekoia_automation.module import Module
 
+from thor_cloud_modules.action_generate_launcher import GenerateLauncherAction
 from thor_cloud_modules.connector import ThorCloudConnector
 from thor_cloud_modules.models import ThorCloudModuleConfiguration
 
@@ -8,4 +9,4 @@ class ThorCloudModule(Module):
     configuration: ThorCloudModuleConfiguration
 
 
-__all__ = ["ThorCloudModule", "ThorCloudConnector"]
+__all__ = ["ThorCloudModule", "ThorCloudConnector", "GenerateLauncherAction"]